requirements.txt
streamlit
requests
gspread
gspread-dataframe
yagmail
pandas
google-auth
